# Faster gradient rules for {numpy,scipy}.linalg.solve