Fixes #1747

The implicit function theorem (via `lax.custom_linear_solve`) lets us _directly_ define gradients for linear solves, in contrast to the current implementations of gradient for `solve` which rely upon differentiating matrix factorization.

In **theory**, JVPs of `cholesky` and `lu` involve the equivalent of ~3 dense matrix-matrix multiplications, which makes them rather expensive: time `O(n**3)`. In contrast, with `custom_linear_solve` we don't need to differentiate the factorization. The JVP and VJP rules for linear solve (for a single right-hand-side vector) now only use matrix-vector products and triangular solves, which is time `O(n**2)`. We should also have reduced memory usage, because we don't need to save any intermediate outputs.

In **practice**, these new gradient rules seem to make linear solves with large arrays ~3x faster, e.g.,

    from functools import partial
    import jax.scipy as jsp
    from jax import lax
    import jax.numpy as np
    import numpy as onp
    import jax

    def loss(solve):
      def f(a, b):
        return solve(a, b).sum()
      return f

    rs = onp.random.RandomState(0)
    N = 500
    K = 1
    a = rs.randn(N, N)
    a = jax.device_put(a.T @ a + 0.1 * np.eye(N))
    b = jax.device_put(rs.randn(N, K))

    # general matrix solve
    grad = jax.jit(jax.grad(loss(np.linalg.solve)))
    grad(a, b).block_until_ready()
    %timeit grad(a, b).block_until_ready()
    # N=500, K=1: 11.4 ms -> 3.63 ms

    # positive definite solve
    grad = jax.jit(jax.grad(loss(partial(jsp.linalg.solve, sym_pos=True))))
    grad(a, b).block_until_ready()
    %timeit grad(a, b).block_until_ready()
    # N=500, K=1: 9.22 ms -> 2.83 ms