@shoyer 2020-02-19
thanks for the review @mattjj !